@@ -3,13 +3,11 @@ plugins {
 }
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 31
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 1
-        versionName '1.0'
+        minSdkVersion 21
+        targetSdkVersion 31
         consumerProguardFiles 'proguard-rules.pro'
     }
 
@@ -19,11 +17,16 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
     implementation 'org.nanohttpd:nanohttpd:2.3.1'
-    implementation 'com.google.code.gson:gson:2.8.5'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.squareup.okhttp3:okhttp:4.9.0'
+    implementation 'com.google.code.gson:gson:2.8.8'
+    implementation 'androidx.core:core:1.7.0'
+    implementation 'com.squareup.okhttp3:okhttp:4.9.1'
 }
